Conflicts:
	README.md
	src/main/com/github/perlundq/yajsync/session/Receiver.java
	src/main/com/github/perlundq/yajsync/session/Sender.java
	src/main/com/github/perlundq/yajsync/ui/YajSyncClient.java
	src/main/com/github/perlundq/yajsync/ui/YajSyncServer.java
	src/main/com/github/perlundq/yajsync/util/Environment.java
	src/main/com/github/perlundq/yajsync/util/Option.java

+++ src/main/com/github/perlundq/yajsync/session/Receiver.java
@@ -1470,7 +1470,7 @@ public class Receiver implements RsyncTask,MessageHandler
         ByteBuffer replicaBuf =
             ByteBuffer.allocate(sizeForChecksumBlock(blockIndex,
                                                      checksumHeader));
-        long fileOffset = blockIndex * checksumHeader.blockLength();
+        long fileOffset = (long) blockIndex * checksumHeader.blockLength();
         replica.position(fileOffset);
         int bytesRead = replica.read(replicaBuf);
         if (replicaBuf.hasRemaining()) {
